+++ app/imports/ui/utilities/HomePageProps.ts
@@ -2,4 +2,5 @@ export interface HomePageProps {
   okItems: JSX.Element[];
   reviewItems: JSX.Element[];
   improveItems: JSX.Element[];
+  role?: string;
 }
